+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -134,7 +134,7 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-19T14:53:43.921Z")
+    created: Optional[datetime] = parse_datetime("2026-01-19T15:09:21.418Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")