+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFBBZ0NB9RK1KS31WQSQ6NTY"
+        "01KFBCVKKA84XHJ4R6KWJBF5F4"
     )
     r"""The unique identifier of the data source"""