+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFBBZ0NFN04MRK322A8X2CQR"
+        "01KFBCVKKDSZY0XMXTFV4381GA"
     )
     r"""The unique identifier of the data source"""