+++ packages/orq-rc/RELEASES.md
@@ -4799,3 +4799,13 @@ Based on:
 - [python v4.2.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.42 - packages/orq-rc
+
+## 2026-01-19 15:11:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.43 - packages/orq-rc